Make Phoenix into header-only library, remove SDL from Renderer
